requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
APScheduler==3.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
emoji==0.6.0
future==0.18.2
hiredis==1.1.0
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
instabot==0.117.0
instagram-private-api==1.6.0
loguru==0.5.3
mock==4.0.3
moviepy==1.0.3
multidict==5.1.0
numpy==1.19.5
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
proglog==0.1.9
py==1.10.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.4
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.12.1
schedule==0.6.0
six==1.15.0
toml==0.10.2
tqdm==4.55.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
win32-setctime==1.0.3
yarl==1.6.3
